+++ packages/core/src/js/integrations/expocontext.ts
@@ -48,10 +48,14 @@ export const expoContextIntegration = (): Integration => {
   }
 
   function addExpoUpdatesContext(event: Event): void {
+    const updatesContext = getExpoUpdatesContextCached();
+
     event.contexts = event.contexts || {};
     event.contexts[OTA_UPDATES_CONTEXT_KEY] = {
-      ...getExpoUpdatesContextCached(),
+      ...updatesContext,
     };
+
+    addExpoUpdatesTags(event, updatesContext);
   }
 
   function getExpoUpdatesContextCached(): ExpoUpdatesContext {
@@ -109,6 +113,20 @@ export function getExpoUpdatesContext(): ExpoUpdatesContext {
     updatesContext.created_at = expoUpdates.createdAt.toISOString();
   }
   return updatesContext;
+}
+
+function addExpoUpdatesTags(event: Event, updatesContext: ExpoUpdatesContext): void {
+  event.tags = event.tags || {};
+
+  if (updatesContext.update_id) {
+    event.tags['expo.updates.update_id'] = updatesContext.update_id;
+  }
+  if (updatesContext.channel) {
+    event.tags['expo.updates.channel'] = updatesContext.channel;
+  }
+  if (updatesContext.runtime_version) {
+    event.tags['expo.updates.runtime_version'] = updatesContext.runtime_version;
+  }
 }
 
 function addExpoGoContext(event: Event): void {