+++ CHANGELOG.md
@@ -24,6 +24,19 @@
 - Adds tags with Expo Updates context variables to make them searchable and filterable ([#5788](https://github.com/getsentry/sentry-react-native/pull/5788))
 - Automatically capture a warning event when Expo Updates performs an emergency launch ([#5794](https://github.com/getsentry/sentry-react-native/pull/5794))
 
+### Fixes
+
+- Defer initial navigation span creation until navigation container is registered ([#5789](https://github.com/getsentry/sentry-react-native/pull/5789))
+
+### Dependencies
+
+- Bump CLI from v3.3.0 to v3.3.1 ([#5793](https://github.com/getsentry/sentry-react-native/pull/5793))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#331)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/3.3.0...3.3.1)
+- Bump JavaScript SDK from v10.42.0 to v10.43.0 ([#5791](https://github.com/getsentry/sentry-react-native/pull/5791))
+  - [changelog](https://github.com/getsentry/sentry-javascript/blob/develop/CHANGELOG.md#10430)
+  - [diff](https://github.com/getsentry/sentry-javascript/compare/10.42.0...10.43.0)
+
 ## 8.3.0
 
 ### Features